refactor(aiops): clean up alert rule draft flow
